IT'S ALIIIIVVE. *ahem* Just gotta test it thoroughly. and refactor some things. and write the UT [ACC-0003](#10)

+++ BudgetBunny/Classes/Views/AddEditAccountTableViewCell.swift
@@ -9,18 +9,6 @@
 import CoreData
 import UIKit
 
-protocol PushViewControllerDelegate: class {
-    func pushCurrencyViewController()
-}
-
-protocol PopViewControllerDelegate: class {
-    func popViewController()
-}
-
-protocol SetAsDefaultDelegate: class {
-    func setAsDefault()
-}
-
 class AddEditAccountTableViewCell: UITableViewCell, UITextFieldDelegate {
 
     @IBOutlet weak var field: UILabel!
@@ -30,11 +18,7 @@ class AddEditAccountTableViewCell: UITableViewCell, UITextFieldDelegate {
     @IBOutlet weak var accountSwitch: UISwitch!
     var model: AddEditAccountCell?
     var fieldMaxLength: Int = 0
-    
-    weak var pushDelegate:PushViewControllerDelegate?
-    weak var popDelegate:PopViewControllerDelegate?
-    weak var setDefaultDelegate:SetAsDefaultDelegate?
-    
+    weak var delegate:AddEditAccountDelegate?
     let constants = ScreenConstants.AddEditAccount.self
     
     override func awakeFromNib() {
@@ -124,7 +108,7 @@ class AddEditAccountTableViewCell: UITableViewCell, UITextFieldDelegate {
             break
             
         case Constants.CellIdentifiers.addAccountChevron:
-            self.pushDelegate?.pushCurrencyViewController()
+            self.delegate?.pushCurrencyViewController()
             break
             
         case Constants.CellIdentifiers.addAccountAction:
@@ -142,39 +126,65 @@ class AddEditAccountTableViewCell: UITableViewCell, UITextFieldDelegate {
     
     func setDefault() {
         // TO-DO: Visual bug when the button is pressed. I wonder if we can make it so that it fades instead...
-        // TO-DO: Bug for both Set Default and Delete buttons: both still execute even when disabled.
-        BunnyUtils.keyExistsForCellSettings(self.model!, key: constants.keyManagedObject) { (object) in
-            let managedObject = object as! NSManagedObject
-            let activeRecord = BunnyModel.init(tableName: ModelConstants.Entities.account)
-            activeRecord.selectAllObjects({ (fetchedObjects) -> Void in
-                // For each element
-                for object in fetchedObjects {
-                    // If the element is the currently selected element, set isDefault to true.
-                    if object == managedObject {
-                        object.setValue(true, forKey: ModelConstants.Account.isDefault)
-                    }
+        BunnyUtils.keyExistsForCellSettings(self.model!, key: constants.keyEnabled, completion: { (object) in
+            let isEnabled = object as! Bool
+            if isEnabled {
+                BunnyUtils.keyExistsForCellSettings(self.model!, key: self.constants.keyManagedObject) { (object) in
+                    let managedObject = object as! NSManagedObject
+                    let activeRecord = BunnyModel.init(tableName: ModelConstants.Entities.account)
+                    activeRecord.selectAllObjects({ (fetchedObjects) -> Void in
+                        // For each element
+                        for object in fetchedObjects {
+                            // If the element is the currently selected element, set isDefault to true.
+                            if object == managedObject {
+                                object.setValue(true, forKey: ModelConstants.Account.isDefault)
+                            }
+                                
+                                // Else, if the element is the previously default account, set isDefault to false.
+                            else if object.valueForKey(ModelConstants.Account.isDefault) as! Bool == true {
+                                object.setValue(false, forKey: ModelConstants.Account.isDefault)
+                            }
+                        }
                         
-                    // Else, if the element is the previously default account, set isDefault to false.
-                    else if object.valueForKey(ModelConstants.Account.isDefault) as! Bool == true {
-                        object.setValue(false, forKey: ModelConstants.Account.isDefault)
-                    }
+                        // Save the model, reload the data, etc.
+                        activeRecord.save()
+                    })
+                    self.delegate?.setAsDefault()
                 }
-                
-                // Save the model, reload the data, etc.
-                activeRecord.save()
-            })
-            self.setDefaultDelegate?.setAsDefault()
-        }
+            }
+        })
     }
     
     func deleteAccount() {
-        // TO-DO: Show alert action -- "This account, and all its transactions, will be deleted. Are you sure?"
-        BunnyUtils.keyExistsForCellSettings(self.model!, key: constants.keyManagedObject) { (object) in
-            let activeRecord = BunnyModel.init(tableName: ModelConstants.Entities.account)
-            activeRecord.deleteObject(object as! NSManagedObject, completion: {
-                self.popDelegate?.popViewController()
-            })
-        }        
+        BunnyUtils.keyExistsForCellSettings(self.model!, key: constants.keyEnabled, completion: { (object) in
+            let isEnabled = object as! Bool
+            if isEnabled {
+                let alertController = UIAlertController.init(
+                    title: BunnyUtils.uncommentedLocalizedString(StringConstants.LABEL_WARNING_DELETE_ACCOUNT_TITLE),
+                    message: BunnyUtils.uncommentedLocalizedString(StringConstants.LABEL_WARNING_DELETE_ACCOUNT_MESSAGE),
+                    preferredStyle: UIAlertControllerStyle.ActionSheet
+                )
+                let deleteAction = UIAlertAction.init(
+                    title: BunnyUtils.uncommentedLocalizedString(StringConstants.BUTTON_DELETE_ACCOUNT),
+                    style: UIAlertActionStyle.Destructive,
+                    handler: { (UIAlertAction) in
+                        let activeRecord = BunnyModel.init(tableName: ModelConstants.Entities.account)
+                        activeRecord.deleteObject(object as! NSManagedObject, completion: {
+                            self.delegate?.popViewController()
+                        })
+                    }
+                )
+                let cancelAction = UIAlertAction.init(
+                    title: BunnyUtils.uncommentedLocalizedString(StringConstants.BUTTON_CANCEL),
+                    style: UIAlertActionStyle.Cancel,
+                    handler: nil
+                )
+                
+                alertController.addAction(deleteAction)
+                alertController.addAction(cancelAction)
+                self.delegate?.presentViewController(alertController)
+            }
+        })
     }
     
     func getValue() -> String {
